feat: Mark/unmark favorite quick action
