+++ components/common/TouchableItemRouter.tsx
@@ -1,4 +1,5 @@
 import { useMarkAsPlayed } from "@/hooks/useMarkAsPlayed";
+import { useFavorite } from "@/hooks/useFavorite";
 import {
   BaseItemDto,
   BaseItemPerson,
@@ -7,7 +8,6 @@ import { useRouter, useSegments } from "expo-router";
 import { PropsWithChildren, useCallback } from "react";
 import { TouchableOpacity, TouchableOpacityProps } from "react-native";
 import { useActionSheet } from "@expo/react-native-action-sheet";
-import { useHaptic } from "@/hooks/useHaptic";
 
 interface Props extends TouchableOpacityProps {
   item: BaseItemDto;
@@ -57,14 +57,14 @@ export const TouchableItemRouter: React.FC<PropsWithChildren<Props>> = ({
   const segments = useSegments();
   const { showActionSheetWithOptions } = useActionSheet();
   const markAsPlayedStatus = useMarkAsPlayed([item]);
-
+  const { isFavorite, toggleFavorite, _} = useFavorite(item);
+  
   const from = segments[2];
 
   const showActionSheet = useCallback(() => {
     if (!(item.Type === "Movie" || item.Type === "Episode")) return;
-
-    const options = ["Mark as Played", "Mark as Not Played", "Cancel"];
-    const cancelButtonIndex = 2;
+    const options = ["Mark as Played", "Mark as Not Played", isFavorite ? "Unmark as Favorite" : "Mark as Favorite", "Cancel"];
+    const cancelButtonIndex = 3;
 
     showActionSheetWithOptions(
       {
@@ -74,14 +74,14 @@ export const TouchableItemRouter: React.FC<PropsWithChildren<Props>> = ({
       async (selectedIndex) => {
         if (selectedIndex === 0) {
           await markAsPlayedStatus(true);
-          // Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
         } else if (selectedIndex === 1) {
           await markAsPlayedStatus(false);
-          // Haptics.notificationAsync(Haptics.NotificationFeedbackType.Success);
+        } else if (selectedIndex === 2) {
+          toggleFavorite()
         }
       }
     );
-  }, [showActionSheetWithOptions, markAsPlayedStatus]);
+  }, [showActionSheetWithOptions, isFavorite, markAsPlayedStatus]);
 
   if (
     from === "(home)" ||
